Using int64 in file.cc. Fixing compilation error in array_utils.cc

+++ internal/ceres/array_utils.cc
@@ -52,7 +52,7 @@ bool IsArrayValid(const int64_t size, const double* x) {
   return true;
 }
 
-int FindInvalidValue(const int64_t size, const double* x) {
+int64_t FindInvalidValue(const int64_t size, const double* x) {
   if (x == nullptr) {
     return size;
   }
